+++ examples/comparison_cantilever_beam.py
@@ -1,4 +1,4 @@
-from Problems.svanberg1987 import CantileverBeam
+from problems.svanberg1987 import CantileverBeam
 from sao.move_limits import Bounds, MoveLimit, MoveLimitFraction, AdaptiveMoveLimit
 from sao.intervening_variables.mma import MMA87A, MMA02, MMA87C
 from sao.intervening_variables import Linear, ConLin, ReciCubed, Reciprocal, Exponential
