+++ examples/topopt/eigenfrequency_clamped_beam.py
@@ -3,12 +3,12 @@ import matplotlib.pyplot as plt
 from matplotlib import colors
 
 
-from Problems.topology_optimization import utils
+from problems.topology_optimization import utils
 from sao.problems import Subproblem
 from sao.move_limits import AdaptiveMoveLimit, Bounds, MoveLimit
 from sao.intervening_variables import Linear, MMA, MixedIntervening
 from sao.approximations import Taylor1
-from Problems.topology_optimization.eigenfrequency import EigenfrequencyClampedBeam
+from problems.topology_optimization.eigenfrequency import EigenfrequencyClampedBeam
 from examples.topopt.optimize import optimize
 
 itercount = 50
