+++ examples/topopt/mechanism_clamped_beam.py
@@ -2,7 +2,7 @@ import numpy as np
 from matplotlib import colors
 from scipy.sparse import coo_matrix
 
-from Problems.topology_optimization import utils
+from problems.topology_optimization import utils
 import matplotlib.pyplot as plt
 
 from sao.problems import Subproblem
@@ -10,7 +10,7 @@ from sao.move_limits import AdaptiveMoveLimit, Bounds, MoveLimit
 from sao.intervening_variables import Linear, MMA, MixedIntervening
 from sao.approximations import Taylor1
 
-from Problems.topology_optimization.mechanism import MechanismClampedBeam
+from problems.topology_optimization.mechanism import MechanismClampedBeam
 
 from examples.topopt.optimize import optimize
