+++ examples/topopt/selfweight_arch.py
@@ -1,15 +1,14 @@
 import numpy as np
 from matplotlib import colors
 
-from Problems.topology_optimization import utils
+from problems.topology_optimization import utils
 import matplotlib.pyplot as plt
 from sao.problems import Subproblem
 from sao.move_limits import AdaptiveMoveLimit, Bounds, MoveLimit
 from sao.intervening_variables import Linear, MMA, MixedIntervening
 from sao.approximations import Taylor1
 
-
-from Problems.topology_optimization.compliance import SelfweightArch
+from problems.topology_optimization.compliance import SelfweightArch
 
 
 from examples.topopt.optimize import optimize
