+++ examples/util/performance_measures.py
