+++ examples/vanderplaats.py
@@ -7,8 +7,8 @@ from sao.move_limits import Bounds, MoveLimit, AdaptiveMoveLimit
 from sao.scaling_strategies import InitialObjectiveScaling, InitialResponseScaling
 from sao.util import Plot
 from sao.solvers import SvanbergIP
-from util.plotter import Plot2, Plot3
-from Problems._nd.VanderplaatsBeam import Vanderplaats
+from examples.util.plotter import Plot2, Plot3
+from problems._nd.VanderplaatsBeam import Vanderplaats
 
 # Set options for logging data: https://www.youtube.com/watch?v=jxmzY9soFXg&ab_channel=CoreySchafer
 logger = logging.getLogger(__name__)
