+++ problems/_1d/Polynomial_1D.py
