+++ problems/_2d/Townsend_func.py
@@ -4,7 +4,7 @@ import math
 from sao.problems.problem import Problem
 
 
-## CLASS: see ReferenceFiles/Problems.pdf by https://en.wikipedia.org/wiki/Test_functions_for_optimization
+## CLASS: see reference_files/problems.pdf by https://en.wikipedia.org/wiki/Test_functions_for_optimization
 class Townsend(Problem):
 
     def __init__(self):
