+++ problems/svanberg1987.py
@@ -1,6 +1,6 @@
 from sao.problems.problem import Problem
 import numpy as np
-from Problems.utils import finite_difference
+from problems.utils import finite_difference
 
 
 class CantileverBeam(Problem):
